api/requirements.txt
aiohttp==3.5.1
aiohttp-graphql==1.0.0
aniso8601==3.0.2
appnope==0.1.0
async-timeout==3.0.1
attrs==18.2.0
backcall==0.1.0
chardet==3.0.4
decorator==4.3.0
graphene==2.1.3
graphql-core==2.1
graphql-relay==0.4.5
graphql-server-core==1.1.1
idna==2.8
idna-ssl==1.1.0
ipdb==0.11
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.2
multidict==4.5.2
numpy==1.15.4
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
promise==2.2.1
prompt-toolkit==2.0.7
ptyprocess==0.6.0
Pygments==2.3.1
pyorbital==1.5.0
pytest-runner==4.2
Rx==1.6.1
scipy==1.2.0
six==1.12.0
traitlets==4.3.2
typing==3.6.6
typing-extensions==3.6.6
wcwidth==0.1.7
yarl==1.3.0
